Root nodes should never be leafs
The previous situation was a bit strange where if you
had not added anything to the tree yet, you would have
two leafs with each zero hits (the 'any' nodes for ipv4
and ipv6).

This situation was confusing so the new situation is that
root nodes will never be returned as leaf nodes.